newwwwwwwww1-main/smart-internship-tracker/requirements.txt
Flask==3.1.0
mysql-connector-python==9.2.0
python-dotenv==1.1.0
bcrypt==4.3.0
PyJWT==2.10.1
flask-socketio==5.5.1
flask-cors==5.0.1
flask-limiter==3.12
